Moves TEMLogger.h to /include
