+++ src/BgcData.cpp
@@ -1,4 +1,4 @@
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 #include "../include/BgcData.h"
 extern src::severity_logger< severity_level > glg;
