+++ src/Climate.cpp
@@ -12,7 +12,7 @@
 #include "../include/timeconst.h"
 
 #include "TEMUtilityFunctions.h"
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
