+++ src/Cohort.cpp
@@ -20,7 +20,7 @@
 
 #include <boost/assign/list_of.hpp> // for 'list_of()'
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 #include "TEMUtilityFunctions.h"
 
 #include "../include/Cohort.h"
