+++ src/CohortData.cpp
@@ -2,7 +2,7 @@
 
 #include "TEMUtilityFunctions.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 
 extern src::severity_logger< severity_level > glg;
