+++ src/CohortLookup.cpp
@@ -7,7 +7,7 @@
 #include <vector> 
 #include <list>
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 #include "TEMUtilityFunctions.h"
