+++ src/Ground.cpp
@@ -18,7 +18,7 @@
 #include <boost/tokenizer.hpp>
 
 #include "../include/Ground.h"
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
