+++ src/Integrator.cpp
@@ -17,7 +17,7 @@
 *
 */
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
