+++ src/Layer.cpp
@@ -3,7 +3,7 @@
  */
 #include "../include/Layer.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 Layer::Layer() {
