+++ src/MineralInfo.cpp
@@ -1,6 +1,6 @@
 #include "../include/MineralInfo.h"
 #include "TEMUtilityFunctions.h" 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
