+++ src/MineralLayer.cpp
@@ -3,7 +3,7 @@
  */
 #include "../include/MineralLayer.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 MineralLayer::MineralLayer(const double & pdz,
