+++ src/ModelData.cpp
@@ -3,7 +3,7 @@
 #include <boost/filesystem.hpp>
 #include "../include/ModelData.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 #include "TEMUtilityFunctions.h"
 
 extern src::severity_logger< severity_level > glg;
