+++ src/OrganicLayer.cpp
@@ -3,7 +3,7 @@
  */
 #include "../include/OrganicLayer.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 OrganicLayer::OrganicLayer(const double & pdz, const int & type) {
