+++ src/ParentLayer.cpp
@@ -3,7 +3,7 @@
  */
 #include "../include/ParentLayer.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 ParentLayer::ParentLayer(const double & thick) {
