+++ src/RestartData.cpp
@@ -25,7 +25,7 @@
 #include "../include/RestartData.h"
 #include "TEMUtilityFunctions.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
