+++ src/Richards.cpp
@@ -1,6 +1,6 @@
 #include "../include/Richards.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 Richards::Richards() {
