+++ src/Runner.cpp
@@ -10,7 +10,7 @@
 #include "../include/Runner.h"
 #include "../include/Cohort.h"
 #include "TEMUtilityFunctions.h"
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 #include "../include/tbc-debug-util.h"
 
 extern src::severity_logger< severity_level > glg;
