+++ src/SoilLayer.cpp
@@ -3,7 +3,7 @@
  */
 #include "../include/SoilLayer.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 SoilLayer::SoilLayer() {
