+++ src/Soil_Bgc.cpp
@@ -22,7 +22,7 @@
  *
  */
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 #include "../include/Soil_Bgc.h"
