+++ src/TEM.cpp
@@ -70,7 +70,7 @@
 #include "../include/timeconst.h"
 #include "../include/ArgHandler.h"
 #include "../include/OutputEstimate.h"
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 #include "TEMUtilityFunctions.h"
 #include "../include/Runner.h"
 #include "../include/RestartData.h"
