+++ src/TemperatureUpdator.cpp
@@ -7,7 +7,7 @@
 
 #include "../include/TemperatureUpdator.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 #include "TEMUtilityFunctions.h"
 
 extern src::severity_logger< severity_level > glg;
