+++ src/Vegetation.cpp
@@ -20,7 +20,7 @@
 
 #include "TEMUtilityFunctions.h"
 
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 Vegetation::Vegetation() {
