+++ src/WildFire.cpp
@@ -27,7 +27,7 @@
 #include "../include/WildFire.h"
 
 #include "TEMUtilityFunctions.h"
-#include "TEMLogger.h"
+#include "../include/TEMLogger.h"
 
 extern src::severity_logger< severity_level > glg;
